Refactor memory management and enhance performance metrics
- Updated `getMemoryFragmentation` to ensure proper type casting for available memory size.
- Introduced `getMemoryLoadPercentage` in `linux.cpp` to calculate and log memory load as a percentage.
- Added corresponding declaration for `getMemoryLoadPercentage` in `linux.hpp`.
- Cleaned up GPIO includes in `gpio.cpp` for better readability.
- Improved network connection handling in `network_manager.cpp` by using a more explicit array initialization.
- Refactored CPU usage calculation in `pidwatcher.cpp` to improve clarity and maintainability, removing unused Windows-specific code.
- Streamlined process CPU usage calculation in `process.cpp`, ensuring consistent variable naming.
- Enhanced address info management in `addr_info.cpp` by using `noexcept` for lambda functions.
- Updated CMake warning messages for ccache installation instructions to improve clarity and accessibility.

+++ atom/sysinfo/memory/common.cpp
@@ -179,7 +179,7 @@ auto getMemoryFragmentation() -> double {
     size_t allocatableSize = 0;
     try {
         constexpr size_t MAX_ALLOC_SIZE = 100 * 1024 * 1024;  // 100 MB
-        const size_t testSize = std::min(MAX_ALLOC_SIZE, available);
+        const size_t testSize = std::min(MAX_ALLOC_SIZE, static_cast<size_t>(available));
         std::vector<char> testAlloc;
         testAlloc.reserve(testSize);
         allocatableSize = testAlloc.capacity();
